### 2021-09-28 #3774 @whedon
[PRE REVIEW]: The jagstargets R package: a reproducible workflow framework for Bayesian data analysis with JAGS
**Submitting author:** @wlandau (<a href="http://orcid.org/0000-0003-1878-3253">William Landau</a>)
**Repository:** <a href="https://github.com/ropensci/jagstargets" target ="_blank">https://github.com/ropensci/jagstargets</a>
**Version:** 0.0.1
**Reviewers:** @AoifeHughes, @sachsmc, @JonasMoss, MaaniBeigy
