### 2021-09-28 f796440
Revert "Allow package to build if JAGS is not installed via lazy-loading"
