+++ R/tar_jags_package.R
@@ -18,6 +18,8 @@
 #' @importFrom posterior summarize_draws
 #' @importFrom purrr map map_dbl map_dfr
 #' @importFrom qs qread
+#' @importFrom R2jags jags jags.parallel
+#' @importFrom rjags load.module
 #' @importFrom rlang expr quo_squash
 #' @importFrom stats rnorm runif
 #' @importFrom targets tar_assert_chr tar_assert_in
