+++ NAMESPACE
@@ -9,6 +9,8 @@ export(tar_jags_rep_draws)
 export(tar_jags_rep_run)
 export(tar_jags_rep_summary)
 export(tar_jags_run)
+importFrom(R2jags,jags)
+importFrom(R2jags,jags.parallel)
 importFrom(coda,mcmc)
 importFrom(coda,mcmc.list)
 importFrom(digest,digest)
@@ -18,6 +20,7 @@ importFrom(purrr,map)
 importFrom(purrr,map_dbl)
 importFrom(purrr,map_dfr)
 importFrom(qs,qread)
+importFrom(rjags,load.module)
 importFrom(rlang,expr)
 importFrom(rlang,quo_squash)
 importFrom(stats,rnorm)
