+++ tests/testthat/test-tar_jags_rep_dic.R
@@ -7,6 +7,8 @@
 # checks that the pipeline is correctly constructed
 # and the output is correctly formatted.
 targets::tar_test("tar_jags_rep_dic()", {
+  skip_if_not_installed("rjags")
+  skip_if_not_installed("R2jags")
   skip_on_cran()
   skip_if_not_installed("dplyr")
   tar_jags_example_file(path = "a.jags")
@@ -127,6 +129,8 @@ targets::tar_test("tar_jags_rep_dic()", {
 })
 
 targets::tar_test("tar_jags_rep_dic() errors correctly if no JAGS file", {
+  skip_if_not_installed("rjags")
+  skip_if_not_installed("R2jags")
   expect_error(
     tar_jags_rep_dic(
       model,
