+++ tests/testthat/test-tar_jags.R
@@ -9,6 +9,8 @@
 targets::tar_test("tar_jags()", {
   skip_on_cran()
   skip_if_not_installed("dplyr")
+  skip_if_not_installed("rjags")
+  skip_if_not_installed("R2jags")
   tar_jags_example_file(path = "a.jags")
   tar_jags_example_file(path = "b.jags")
   targets::tar_script({
@@ -153,6 +155,8 @@ targets::tar_test("tar_jags()", {
 targets::tar_test("tar_jags() can produced user-defined MCMC summaries", {
   skip_on_cran()
   skip_if_not_installed("dplyr")
+  skip_if_not_installed("rjags")
+  skip_if_not_installed("R2jags")
   tar_jags_example_file(path = "a.jags")
   tar_jags_example_file(path = "b.jags")
   targets::tar_script({
