### 2021-09-28 f85bc20
Test without rjags or R2jags
