+++ man/tar_jags.Rd
@@ -55,8 +55,9 @@ generated by the code in \code{data}. If you supply an unnamed vector,
 as target name suffixes. If \code{jags_files} is a named vector,
 the suffixed will come from \code{names(jags_files)}.}
 
-\item{parameters.to.save}{character vector of the names of the
-    parameters to save which should be monitored.}
+\item{parameters.to.save}{Model parameters to save, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
 \item{data}{Code to generate the \code{data} list for the JAGS model.
 Optionally include a \code{.join_data} element to join parts of the data
@@ -71,37 +72,41 @@ on the \code{CmdStanFit} object.}
 \code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
 on the \code{CmdStanFit} object.}
 
-\item{n.cluster}{number of clusters to use to run parallel chains.
-    Default equals n.chains.}
+\item{n.cluster}{Number of parallel processes, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{n.chains}{number of Markov chains (default: 3)}
+\item{n.chains}{Number of MCMC chains, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{n.iter}{number of total iterations per chain (including burn in;
-    default: 2000)}
+\item{n.iter}{Number if iterations (including warmup), passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{n.burnin}{length of burn in, i.e. number of iterations to
-    discard at the beginning. Default is \code{n.iter/2}, that is,
-    discarding the first half of the simulations. If n.burnin is 0,
-    \code{jags()} will run 100 iterations for adaption.}
+\item{n.burnin}{Number of warmup iterations, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{n.thin}{thinning rate. Must be a positive integer.  Set
-    \code{n.thin} > 1 to save memory and computation time if
-    \code{n.iter} is large.  Default is \code{max(1, floor(n.chains *
-    (n.iter-n.burnin) / 1000))} which will only thin if there are at
-    least 2000 simulations.}
+\item{n.thin}{Thinning interval, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{jags.module}{the vector of jags modules to be loaded.  Default are \dQuote{glm} and \dQuote{dic}. Input NULL if you don't want to load any jags module.}
+\item{jags.module}{Character vector of JAGS modules to load, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{inits}{a list with \code{n.chains} elements; each element of the
-    list is itself a list of starting values for the \code{BUGS} model,
-    \emph{or} a function creating (possibly random) initial values. If inits is
-    \code{NULL}, \code{JAGS} will generate initial values for parameters.}
+\item{inits}{Initial values of model parameters, passed to
+\code{R2jags::jags()}. See the argument documentation of the
+\code{R2jags::jags()} help file for details.}
 
-\item{RNGname}{the name for random number generator used in JAGS. There are four RNGS
-  supplied by the base moduale in JAGS: \code{Wichmann-Hill}, \code{Marsaglia-Multicarry},
-  \code{Super-Duper}, \code{Mersenne-Twister}}
[…]
