### 2021-09-28 9e38793
Do not inherit params from R2jags or rjags
