+++ NAMESPACE
@@ -1,5 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
+export(tar_jags)
+export(tar_jags_df)
+export(tar_jags_example_data)
+export(tar_jags_example_file)
+export(tar_jags_rep_dic)
+export(tar_jags_rep_draws)
+export(tar_jags_rep_run)
+export(tar_jags_rep_summary)
+export(tar_jags_run)
 importFrom(coda,mcmc)
 importFrom(coda,mcmc.list)
 importFrom(digest,digest)

[…]
