+++ DESCRIPTION
@@ -44,6 +44,8 @@ Imports:
   posterior (>= 1.0.1),
   purrr (>= 0.3.4),
   qs (>= 0.23.2),
+  R2jags (>= 0.6.1),
+  rjags (>= 4.10),
   rlang (>= 0.4.10),
   stats,
   targets (>= 0.6.0),
@@ -57,8 +59,6 @@ Suggests:
   fs (>= 1.5.0),
   knitr (>= 1.30),
   R.utils (>= 2.10.1),
-  R2jags (>= 0.6.1),
-  rjags (>= 4.10),
   rmarkdown (>= 2.3),
   testthat (>= 3.0.0),
   tidyr (>= 1.1.2),
