### 2021-09-30 #3774 @galessiorob
👋 Hi @wlandau

I am in the process of assigning reviewers to this submission, if you have any suggestions please let me know and feel free to @ mention them here.

Basic checks to go over before we formally start the review:

 - [ ] There’s a general description of the software
           Needs expanding
 - [ ] The software is within the scope as research software
           Potentially to be expanded to denote research application and use cases
 - [ ] OSI-approved license: GNU General Public License (GPL)
