### 2021-09-30 #3774 @wlandau-lilly
Hi @galessiorob, thanks for reaching out.

> OSI-approved license: GNU General Public License (GPL)

Yes, `jagstargets` is currently under MIT. Does the software need to be switched to GPL for the JOSS manuscript to be considered?
