### 2021-10-05 #3774 @galessiorob
@wlandau MIT license is just fine 👍

I've emailed several volunteers that I think are a good match for your paper, but feel free to recommend others by mentioning them here. Thanks!
