+++ inst/paper.bib
@@ -124,3 +124,12 @@
   title = {The stantargets {R} package: a workflow framework for efficient reproducible {S}tan-powered {B}ayesian data analysis pipelines},
   journal = {Journal of Open Source Software}
 }
+
+@misc{
+  carpenter2017,
+  author = {Bob Carpenter},
+  title = {{Bayesian Posteriors are Calibrated by Definition}},
+  year = {2017},
+  url = {https://statmodeling.stat.columbia.edu/2017/04/12/bayesian-posteriors-calibrated/}
+}
+
