+++ inst/paper.md
@@ -28,8 +28,8 @@ Bayesian statisticians regularly experiment with models to refine, compare, and 
 
 # Summary
 
-The [`jagstargets`](https://docs.ropensci.org/jagstargets/) R package [@jagstargets] is a pipeline toolkit for Bayesian data analysis with JAGS. It automatically detects the dependency relationships among the results, models, and datasets, it watches these computational steps for changes, and it deploys the work with optional distributed computing while skipping the tasks that are already synchronized with their upstream dependencies. This automated selective skipping reduces the runtime of successive experiments, and when the entire workflow is up to date, the user has tangible evidence of reproducibility.
+The [`jagstargets`](https://docs.ropensci.org/jagstargets/) R package [@jagstargets] is a workflow toolkit for Bayesian data analysis with JAGS. It helps users express a Bayesian statistical modeling exercise as a formal pipeline with dedicated steps for data generation, analysis, and summarization. Pipelines can be customized for popular use cases: for example, the analysis of a single dataset using multiple alternative models, or a large simulation study to test that a model is implemented correctly [@carpenter2017]. These pipelines, which can be visualized and executed using [`targets`](https://docs.ropensci.org/targets/) [@targets], support key features that increase the efficiency and reproducibility of Bayesian workflows. The steps of a pipeline are automatically orchestrated using optional distributed computing, and up-to-date tasks are automatically skipped if the upstream code and data did not change since the last run. Thus, researchers can quickly iterate on Bayesian workflows while maintaining agreement between the results and the underlying code, models, and datasets.
 
-The [`jagstargets`](https://docs.ropensci.org/jagstargets/) package is a bridge between packages [`R2jags`](https://github.com/suyusung/R2jags) [@R2jags] and [`targets`](https://docs.ropensci.org/targets/) [@targets]. [`targets`](https://docs.ropensci.org/targets/) is a general-purpose pipeline toolkit for reproducible research and high-performance computing, and it is not specific to Bayesian data analysis. [`jagstargets`](https://docs.ropensci.org/jagstargets/) leverages the existing capabilities of [`targets`](https://docs.ropensci.org/targets/) to more easily and concisely express computational pipelines for Bayesian statisticians, from single analyses or large-scale simulation studies. [`jagstargets`](https://docs.ropensci.org/targets/) is similar to the [`stantargets`](https://docs.ropensci.org/stantargets/) R package [@stantargets], the latter of which streamlines pipeline construction of Bayesian data analysis pipelines with Stan [@stan].
+The [`jagstargets`](https://docs.ropensci.org/jagstargets/) package is a bridge between packages [`R2jags`](https://github.com/suyusung/R2jags) [@R2jags] and [`targets`](https://docs.ropensci.org/targets/). [`targets`](https://docs.ropensci.org/targets/) is a general-purpose pipeline toolkit for reproducible research and high-performance computing, and it is not specific to Bayesian data analysis. [`jagstargets`](https://docs.ropensci.org/jagstargets/) leverages the existing capabilities of [`targets`](https://docs.ropensci.org/targets/) to more easily and concisely express computational pipelines for Bayesian statisticians, from single analyses or large-scale simulation studies. [`jagstargets`](https://docs.ropensci.org/targets/) is similar to the [`stantargets`](https://docs.ropensci.org/stantargets/) R package [@stantargets], the latter of which streamlines pipeline construction of Bayesian data analysis pipelines with Stan [@stan].
 
 # References
