### 2021-11-02 #3877 @whedon
[REVIEW]: The jagstargets R package: a reproducible workflow framework for Bayesian data analysis with JAGS
**Submitting author:** @wlandau (<a href="http://orcid.org/0000-0003-1878-3253">William Landau</a>)
**Repository:** <a href="https://github.com/ropensci/jagstargets" target ="_blank">https://github.com/ropensci/jagstargets</a>
**Version:** v1.0.0

@AoifeHughes & @sachsmc & @JonasMoss & @MaaniBeigy, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
