### 2021-11-04 ropensci/jagstargets#25 @AoifeHughes
JOSS Reviewer Comments
Hi there,

Re: [joss review here](https://github.com/openjournals/joss-reviews/issues/3877)

Very happy to be reviewing this package, off the bat it looks great and like we'll have no problems. But, just to do my job properly, I will be very nitpicky. I promise that if any comments read bluntly that it's just for brevity :)

Many of these comments could probably be ignored, but I am giving a "blind" attempt to use the package and noting any issues one may encounter.

- Installing from remotes, maybe worth pointing out that `install.packages("remotes")` is needed?
- License etc. is fine, I wonder if the naming of the authors is okay, usually papers are "Surname et al." rather than "company".
- The maths in https://docs.ropensci.org/jagstargets/articles/introduction.html doesn't seem to display the model correctly at the top, it renders as plain text for me

I'm having some issues with rjags, so reinstalling R (this is a me problem). Will Continue review when I have this sorted...
