### 2021-11-04 ropensci/jagstargets#25 @wlandau
Thanks for taking the time to review, @AoifeHughes. To address your comments:

* 263c50cebe6408efa61df03d3e32a901ad67d418 adds `install.packages("remotes")` to the README.
* For the manuscript, please let me know if the YAML front matter should be changed to correctly state my affiliation. For the package itself, I agree the conventions may seem strange. I have been trying to follow the general guidance at https://stackoverflow.com/questions/40704920/listing-a-company-as-the-copyright-holder-of-an-r-package and http://r-pkgs.had.co.nz/description.html, and the practices of popular packages, e.g. https://github.com/tidyverse/dplyr.
* Unfortunately, I cannot locally reproduce the issue rendering math formulas, and the site is rendered through [r-universe](https://ropensci.org/r-universe/). Hopefully the last commit will fix any transient problems. If not, I can just use plain text.
