### 2021-11-04 ropensci/jagstargets#25 @wlandau
@jeroen, do you know why https://docs.ropensci.org/jagstargets/articles/introduction.html has trouble displaying [this LaTeX code](https://github.com/ropensci/jagstargets/blob/263c50cebe6408efa61df03d3e32a901ad67d418/vignettes/introduction.Rmd#L38-L43)? Is there an issue MathJax in r-universe? The vignette renders just fine on my local machine.
