### 2021-11-04 ropensci/jagstargets#25 @wlandau
This is what I see in Safari:

![Screen Shot 2021-11-04 at 10 20 52 AM](https://user-images.githubusercontent.com/1580860/140330740-b0def720-03d8-4efa-89ad-3ec123c2d6b0.png)
