### 2021-11-04 ropensci/jagstargets#25 @jeroen
IIRC, our theme made mathjax opt-in. Can you add `mathjax: true` to your `_pkgdown.yml`?
