### 2021-11-05 ropensci/jagstargets#26 @sachsmc
JOSS reviewer comments
Hi @wlandau ,

Thanks for the nice package, it is good timing for me to test out as I am currently working on a project using jags. Here come a few minor suggestions on the write up for JOSS:

- [ ] First sentence: suggest rephrasing as "Researchers(/statisticians/people) performing Bayesian analyses regularly experiment with models to ..."
- [ ] Statement of need, general comment: I like what you have so far, but consider going further by additionally explaining something like: "the targets package helps researchers manage the complexity of multiple models, data, etc., but it lacks specific features for dealing with models specified in jags ... "
- [ ] Last paragraph line 33: "Bayesian statisticians" -> "Bayesian data analysis"
- [ ] The "State of the field" description is lacking. If I didn't want to use jagstargets (but still jags), what are my options? I suppose I could write some functional wrappers around the rjags calls and pass them to targets, what would be the disadvantage of doing that aside from the effort?

In general, I suggest avoiding the phrase "Bayesian statisticians" in the documentation as it can be divisive. It is on in the first paragraph of the README as well. It is nit-picky, but one need not identify as a Bayesian statistician to do Bayesian data analyses. Actually now that I think about it you don't even need to be a statistician.

Best regards,

Michael
