### 2021-11-05 2c20af0
Elaborate on the benefits of jagstargets over just targets
