### 2021-11-09 #3877 @galessiorob
@AoifeHughes, thanks so much for your review thus far, I appreciate you taking the time to document the issues you found so thoroughly.

@sachsmc, thanks so much for completing your checklist! Were you able to install and run the package without issues? If you have any additional feedback for @wlandau feel free to post them here or in an individual issue.

@JonasMoss @lauravana @MaaniBeigy, mind updating us on your timelines to get to your review, please? Let me know if you have any questions or are blocked in any way. Cheers!
