### 2021-11-09 #3877 @sachsmc
@galessiorob Yes, I could install and run without any issues, but I already had jags and rjags set up. I had some feedback here (https://github.com/ropensci/jagstargets/issues/26) on the write up that were addressed by the author.
