### 2021-11-17 ropensci/jagstargets#25 @AoifeHughes
Up and running again:

- in the Readme file  you do provide an example script under usage, however this is incomplete without the accompanying model which is specified here: https://docs.ropensci.org/jagstargets/articles/introduction.html would it be useful to just include it on the readme also?

Again, sorry a small comment, everything else is looking well and I think I can finished the review shortly.
