### 2021-11-17 #3877 @AoifeHughes
> > 👋 @AoifeHughes, please update us on how your review is going (this is an automated reminder).
>
> I am on it, don't worry :)

Fairly sure I'm done now, I only had a few minor comments left that are more stylistic than required. I think this is an excellently documented and written software package and recommend for publication.
