+++ README.Rmd
@@ -57,7 +57,24 @@ remotes::install_github("ropensci/jagstargets")
 
 ## Usage
 
-First, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate JAGS data, and lists a pipeline of targets. The target list can call target factories like [`tar_jags()`](https://docs.ropensci.org/jagstargets/reference/tar_jags.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html). The following minimal example is simple enough to contain entirely within the `_targets.R` file, but for larger projects, you may wish to store functions in separate files as in the [`targets-stan`](https://github.com/wlandau/targets-stan) example.
+Begin with one or more models: for example, the simple regression model below with response variable $y$ and covariate $x$.
+
+<center>
+<img src="./man/figures/model.gif">
+</center>
+
+Next, write a JAGS model file for each model like the `model.jags` file below.
+
+```jags
+model {
+  for (i in 1:n) {
+    y[i] ~ dnorm(x[i] * beta, 1)
+  }
+  beta ~ dnorm(0, 1)
+}
+```
+
+To begin a reproducible analysis pipeline with this model, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate JAGS data, and lists a pipeline of targets. The target list can call target factories like [`tar_jags()`](https://docs.ropensci.org/jagstargets/reference/tar_jags.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html). The following minimal example is simple enough to contain entirely within the `_targets.R` file, but for larger projects, you may wish to store functions in separate files as in the [`targets-stan`](https://github.com/wlandau/targets-stan) example.
 
 ```{r, eval = FALSE}
 # _targets.R
@@ -74,7 +91,7 @@ generate_data <- function() {
 list(
   tar_jags(
     example,
-    jags_files = "x.jags", # You provide this file.
+    jags_files = "model.jags", # You provide this file.
     parameters.to.save = "beta",
     data = generate_data()
   )
