### 2021-11-18 ropensci/jagstargets#25 @wlandau
Sure, added the model to the README in 9ca970631254f359be2ac008b0e4f766fd2dd6c4.
