### 2021-11-30 ropensci/jagstargets#25 @kthyng
@AoifeHughes is this issue completed, in terms of your review for JOSS?
