### 2021-11-30 #3877 @kthyng
Hi @wlandau! I'm here to help wrap up your submission. I see that this was already reviewed by rOpenSci — perhaps your submission is now extra-reviewed! A few notes to start:
1. Please update the metadata in your Zenodo archive so the title and author list exactly match your JOSS paper.
2. Is your software package in beta mode? I notice that the version is 0.0.1, which would imply it's very early in development. Are you planning up bump up the version number?
