### 2021-11-30 #3877 @kthyng
@wlandau Also please modify your .bib file to improve the capitalization in your reference list. You can put {} around characters to preserve capitalization. For example (but please look carefully through all references to check for more), R and Bayesian aren't capitalized.
