+++ inst/paper.bib
@@ -3,13 +3,13 @@
   doi = {10.21105/joss.02959},
   url = {https://doi.org/10.21105/joss.02959},
   year = {2021},
-  publisher = {The Open Journal},
+  publisher = {{The Open Journal}},
   volume = {6},
   number = {57},
   pages = {2959},
   author = {William Michael Landau},
-  title = {The targets R package: a dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing},
-  journal = {Journal of Open Source Software}
+  title = {{The targets R package: a dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing}},
+  journal = {{Journal of Open Source Software}}
 }
 
 @book{
@@ -26,7 +26,7 @@
   title = {{memoise: Memoisation of Functions}},
   author = {Hadley Wickham and Jim Hester and Kirill Müller and Daniel Cook},
   year = {2017},
-  note = {R package version 1.1.0},
+  note = {{R package version 1.1.0}},
   howpublished = {https://CRAN.R-project.org/package=memoise}
 }
 
@@ -41,7 +41,7 @@
 
 @misc{
   bayesworkflow,
-  title = {Bayesian Workflow}, 
+  title = {{Bayesian Workflow}}, 
   author = {Andrew Gelman and Aki Vehtari and Daniel Simpson and Charles C. Margossian and Bob Carpenter and Yuling Yao and Lauren Kennedy and Jonah Gabry and Paul-Christian Bürkner and Martin Modrák},
   year = {2020},
   eprint = {2011.01808},
@@ -53,22 +53,22 @@
   bda3,
   author         = {Andrew Gelman and John B. Carlin and Hal S. Stern and David B. Dunson and Aki Vehtari and Donald B. Rubin},
   editor         = {Francesca Dominici and Julian J. Faraway and Martin Tanner and Jim idek},
-  title          = {Bayesian Data Analysis},
+  title          = {{Bayesian Data Analysis}},
   edition = {Third},
-  publisher      = {CRC Press},
+  publisher      = {{CRC Press}},
   year           = {2014}
 }
 
 @Manual{
   jagstargets,
-  title = {jagstargets: Targets for {JAGS} Workflows},
+  title = {{jagstargets: Targets for JAGS Workflows}},
   author = {William Michael Landau},
   note = {https://docs.ropensci.org/jagstargets/, https://github.com/ropensci/jagstargets},
 }
 
 @Manual{
   cmdstanr,
-  title = {cmdstanr: {R} Interface to '{C}md{S}tan'},
+  title = {{cmdstanr: R Interface to 'CmdStan'}},
   author = {Jonah Gabry and Rok Češnovar},
   year = {2020},
   note = {https://mc-stan.org/cmdstanr, https://discourse.mc-stan.org}
@@ -92,9 +92,9 @@
 
 @inproceedings{
   jags,
-  title={JAGS: A program for analysis of Bayesian graphical models using Gibbs sampling},
+  title={{JAGS: A program for analysis of Bayesian graphical models using Gibbs sampling}},
   author={Plummer, Martyn and others},
-  booktitle={Proceedings of the 3rd international workshop on distributed statistical computing},
+  booktitle={{Proceedings of the 3rd international workshop on distributed statistical computing}},
   volume={124},
   number={125.10},
   pages={1--10},
@@ -104,10 +104,10 @@
 
 @Manual{
   R2jags,
-  title = {R2jags: Using {R} to Run '{JAGS}'},
+  title = {{R2jags: Using {R} to Run '{JAGS}'}},
   author = {Yu-Sung Su and Masanao Yajima,},
   year = {2020},
-  note = {R package version 0.6-1},
[…]
