### 2021-11-30 #3877 @wlandau
Hi @kthyng!

> Please update the metadata in your Zenodo archive so the title and author list exactly match your JOSS paper.

I added a custom `.zenodo.json` to the repo to ensure Zenodo generates proper metadata.

> Is your software package in beta mode? I notice that the version is 0.0.1, which would imply it's very early in development. Are you planning up bump up the version number?

It is ready for serious use and back-compatibility is a priority, so I bumped the version to 1.0.0.

> Also please modify your .bib file to improve the capitalization in your reference list.

Should be fixed now.
